requirements.txt
fastapi
uvicorn
tortoise-orm
pydantic
